requirements.txt
torch
torchvision
numpy
Pillow
scikit-learn
streamlit
clip-anytorch

